refactor(ui): extract hidden input field into standalone component, move utils that are only used there into it
