+++ ui/src/routes/register/avatar/+page.svelte
@@ -9,8 +9,8 @@
   import { t } from "$lib/translations";
   import { RelayClient } from "$store/RelayClient";
   import { ProfileCreateStore } from "$store/ProfileCreateStore";
-  import { handleFileChange } from "$lib/utils";
   import toast from "svelte-french-toast";
+  import HiddenFileInput from "$lib/HiddenFileInput.svelte";
 
   const relayClientContext: { getClient: () => RelayClient } = getContext("relayClient");
   let relayClient = relayClientContext.getClient();
@@ -45,17 +45,12 @@
 <div class="flex grow flex-col items-center justify-center">
   <h1 class="h1 mb-10">{$t("common.select_an_avatar")}</h1>
 
-  <!-- Hidden file input -->
-  <input
-    type="file"
+  <HiddenFileInput
     accept="image/*"
     id="avatarInput"
-    class="hidden"
-    on:change={(event) =>
-      handleFileChange(event, (imageData) => {
-        ProfileCreateStore.update((current) => {
-          return { firstName, lastName, avatar: imageData };
-        });
+    on:change={(e) =>
+      ProfileCreateStore.update((current) => {
+        return { firstName, lastName, avatar: e.detail };
       })}
   />
